refactor(layout): move Navbar and Footer components to layout directory
