+++ frontend/app/components/layout/Navbar.tsx
